requirements.txt
sentence-transformers==3.4.1
faiss-cpu==1.10.0
pypdf==5.4.0
ollama==0.4.7
numpy==2.2.4
streamlit==1.44.1
python-dotenv==1.1.0
rank-bm25==0.2.2            
unstructured>=0.12.0             
beautifulsoup4>=4.12.0           
requests>=2.31.0
pandas>=2.0.0